feat: Support aliases in dummy connector (#5)
Aliases are different usernames which are associated with the same
backend account. For instance, a user might have two aliases on their
account, `user@example.com` and `alias@example.com`. In this case, their
IMAP client would be able to login with either of these addresses as the
username and it would connect to the same backend IMAP account.

This feature allows Bridge to support Proton accounts with multiple
addresses.