feat(GODT-1619): Per connection metadata storage
This patch adds features to the gluon so that we can store per
connection metadata. This is required to implement the IMAP ID
extension, where we need to store IMAP client identifiers for each of
the incoming connections.

To achieve this we queue operations onto the queue which create, delete
and store values in this storage space. This guarantees the operations
are in sync with any potential future IMAP commands.

This patch also updates the data written to disk to include the metadate
storage in order for serialized commands that haven't been execute to
function correctly.

This patch also addresses a bug in pchan.go related to concurrent access
after queue closing (thanks @James).

Finally this patch also address some issues in pchan reported by
golangcli-run.