feat(GODT-1619): IMAP ID Extension (RFC 2917)
This patch introduces the IMAP ID extension to Gluon. The ID information
is stored on the session so that it can be communicated to the client
if this information arrives before the client has logged in.

The connector can then later access this information from the context
using the `GetIMAPIDFromContext()` function.

To verify this in the unit tests, we now record the last IMAP ID
information during the APPEND command. This was the easiest way to
verify this data was set correctly, since it always requires
communication with the connector.

Finally to support meaning full IMAP ID replies, this patch also
introduces the `VersionInfo` type which can be specified when creating
a new `Server` instance. This information is then later used to populate
the response for this command.

W