fix(GODT-1597): Handle permanently deleted messages
This patch ensures that messages that have been marked as deleted by the
connector are evicted from the store and the user database. To do this,
the messages table has new field named `Deleted` which tracks whether
the connector has deleted this message.

On user creation we erase all messages from the database and the store
that are marked as such. On client logout we attempt to delete all
marked messages that are not in use by other sessions which share the
same user.

This patch also adds `Server.RemoveWatcher()` in order to properly clean
up existing watchers that may have been added.

Finally this patch extends the store interface with `Delete()` and the
respective implementations for the on disk and memory storage.