fix(GODT-1621): Fix race condition with temporary IDs
Currently it is possible to have temporary IDs for mailboxes and
messages to become stale. It is possible, due to concurrency, that a
session pushes an operation which still uses a temporary ID that is no
longer valid.

Rather than locking the entire queue and updating all the remaining
entries, we now maintain two separate translation tables which are
evicted via separate operations.

This reduces the amount of time the queue spends in a locking state and
also guarantees that by the time the removal of the temporary ID is
processed by the operation queue no other operations exist that still
refer to this temporary ID.