fix: Integrate goroutine leak detector and fix reports
This patch integrates the goleak goroutine leak detector which is run
with each server instantiation in the unit tests. We also address the
currently detected goroutines that did not exit properly.

Added `Close()` to `Connector` to ensure the update channels can be
closed properly.

There is currently an issue with double invocation of `Connector.Close()`
which will be addressed with GODT-1647.

Finally we also ensure we wait that the goroutine launched with
`user.forward()` finishes execution before we return. Since we can now
properly wait on this, we no longer require a dedicated stop channel to
terminate the goroutine launched in `newUser()`.