test(GODT-1615): Add more tests to trigger SQL error & Improve fix
The previous pagination fix was not the most optimal possible. After
consulting with a database expert, if we sort the results by the row ID
and use a clause where the row ID is greater to the offset we are
currently evaluating, we can get better DB performance.

This patch also adds more tests to find other potential locations which
may trigger this error.

Finally we also fixed the error present in the listing of mailboxes.