chore: Cleanup of database operations
Reduce code duplication and makes the easier to locate and modify. All
database operations now reside in `db_*.go` files and are prefixed with
`DB`.

This patch also applies some of the best practices described in
https://entgo.io/docs/transactions/. All operations which are readonly
operate on a `ent.Client` type rather than `ent.Tx`. Now it is possible
to use these with and without a transaction.