feat(GODT-1719): Gluon benchmark framework
This patch adds gluon_bench in order to enable use to write benchmarks
that run gluon in the same process as the benchmark or externally. The
support for external processes will be added in a future patch.

The benchmark tool supports producing a report to stdout or a json file.
New reporters can be added as needed.

At the moment, there is only one benchmark which measures the time to
append 1000 message to a mailbox. More benchmark will follow in the
future.

Additionally, the tool also supports running multiple benchmarks for a
configurable number of consecutive times with or without persistent
state.

Finally, this patch also extends Gluon with support to profile the
execution of individual commands inside Gluon. This information is
collected by the benchmark tool to produce its reports.