Fix: Add uniqueness constraint to mailbox names
Currently in the Dummy connector, we create a new default INBOX mailbox
on every `Sync`. While this works fine for short lived tests, it causes
issues for cases where we need to run the same state multiple times. In
these cases certain queries fail as the name of the INBOX is not unique.
We have hard coded the value of the INBOX mailbox to a `0` in order to
guarantee that it remains unique on every run.

Finally a constraint has been added to the database table to catch this
error in the future.