feat(GODT-1917): Bechmark interface and server interface.
This patch introduces the `benchmark.Benchmark` interface to enable
benchmark to setup and tear down before actually running their checks.
For instance, this will be used by fetch to determine the number of
messages present in the mailbox.

This patch also introduces the `server.Server` interface to enable the
benchmark to be run against an external server. Current implementations
include a local implementation (run in process) and remote server which
the benchmark doesn't start, but can still connect to.