feat(GODT-1917): Add Fetch benchmark
This patch adds a benchmark for fetch requests. It supports the
following options:

* Read only toggle
* Fetching all messages
* Fetching a random generated sequence sets of messages
* Fetching a sequence sets defined in file
* Amount of messages to fetch.

This patch also includes some small tweaks to flag location so that
general flags such as the mailbox name and the number of clients to use
for the benchmark can be shared with all benchmark, if required.

Finally we also disable gosec lints for the benchmarks folder as this is
not required.