feat(GODT-1611): Handle optional charset in SEARCH command
The IMAP SEARCH command supports an optional charset. If this is
provided, search terms should be decoded before being compared with
search targets.

In addition, RFC3501 says that search targets should be decoded before
comparison, but this is out of scope for this PR, as it requires parsing
and decoding of messages and addresses.