feat(GODT-1728): Store benchmarks
Add a collection of benchmarks for the message storage cache. All
operations can be benchmarked.

The number of workers set via `store-workers` will determine the amount
of the parallel requests used by the benchmark and `store-item-size`
controls the size of the data written to the store instance.

Finally the number of items to write to storage is controlled via the
`store-item-count flag`.