feat(GODT-1570): Badger store implementation
Add implementation of Badger (https://github.com/dgraph-io/badger) as
an alternative cache implementation. This will become useful later as
part of GODT-1642 where we need to ensure that cache modifications can
be rolled back with failed db modifications and vice-versa.

The unit tests have been updated to use this new implementation and it
is also available to gluon-bench using the `-store=badger` option.

Finally the `StoreBuilder` has been updated to accept a slice of bytes
rather than a string as the encryption key.