feat(GODT-1720, GODT-1722): Implement Database Wrapper
Implements database as discussed in the RFC. It makes distinctions
between read and write operations.

Currently, due to the use of SQLite we only support only one concurrent
writer. This is enforced by a RWLock around the `Read` and `Write`
functions.

Updates to rest of the codebase will follow in future patches.