feat(GODT-1724): Guard database access with an RWLock
Ensure all the database queries are accessed through the scope of RWLock
to avoid the infamous "database locked" error and to allow more parallel
read only requests from clients.

The RWLock guarantees that only one goroutine/thread can write to the
datbase, therefore ensuring we never trigger the limitations of SQLite
DB which causes the error.

The removal of the global TX Lock now enables clients to perform more
read only request in parallel.

Going forward it is expected all database access to be accessed via the
`DB` type.