fix(GODT-1585): Work around race condition in go-imap client
Rework the TestReceptionOnIdle to work around a race condition in
go-imap client library. The race condition happens due the go-imap
client library passing a pointer to `MailboxUpdate` and then re-using it
for the next update.

At the same time a user is processing updates, go-imap will overwrite
the data at the same memory address.