fix(GODT-1585): Fix race conditions in snapshot access
Ensure snapshot access is protected by RWLock as there are cases where
we are closing the snapshot while it is being read when processing
Connector updates.

All access to the snapshot should go through the `snapshotWrapper` type.
Due to limitations of the Go generic system, standalone functions
`snapshotRead` and `snapshotWrite` need to be used.