fix(GODT-1642): Storage database transactions
Add transaction control to the store interface and update existing write
methods that interact with the store to use a transaction.

Since we can't guarantee that the database and the store transactions
commit at the same time, we will commit the store first and then the
database as we can afford to have stale data in the store, but not in
the IMAP state.

The other store implementations have been removed as they don't provide
any transaction mechanism and haven't been used in a while.