fix(GODT-1757): Recent flag behavior and snapshot db reads
This patch removes the need for the snapshots to read new messages from
the database when they arrive via an `exists` responder.

We have also updated the behavior of the `Recent` flag to match
dovecot's implementation. If a client append a message to a selected
mailbox only that client will receive the recent flags. If a client
appends a message to a non-selected mailbox, the first state which
receives the update will apply the recent flag.

To achieve the latter, we added a new update `ExistsStateUpdate` which
enforces the logic described above in a thread safe manner.