fix(GODT-1575,GODT-1757): Untagged responses behavior
This patch fixes the behavior of untagged responses to match dovecot.
All untagged responses are now sent at the end of the command before the
closing tag is sent and no longer before and after. Some of the unit
tests had to be updated to match this new behavior.

This patch also fixes fetch so that it applies the seen flag to the
snapshot on which fetch is operating. This is now possible as snapshots
aren't shared between sessions anymore.