fix(GODT-1757): Flag updates can only be applied through responders
We should only apply flag updates through responders as it possible to
run into a sequence of events where an exist responder hasn't been
processed yet. Attempting to apply the flag changes during a state
update will result in an error or no op.

The error was not reproducible at the moment since the state filter
would fail to queue the update as the message did not exist yet.

Finally the whole issue is currently masked since we get the new
message for the snap shot directly from the database. I ran into this
issue with local changes where this was no longer the case.