Refactor: Strong types for UID UIDValidity and IMAP Sequence (#104)
* refactor: Strong type StateID

* refactor(GODT-1757): Strong Type UID and UIDValidaty

This patch also ensures that these two types are represented by uint32
which matches the spec's requirements.

* refactor(GODT-1757): Strong Type IMAP Sequence

This patch also ensures that these this type is represented by uint32
which matches the spec's requirements.

* test: Fix unit test with missing imap.UID types

* refactor: StateIDs are generated by the state package

* refactor: Remove unecessary changes

Some changes were made to correctly handle the case when a user deletes
a selected mailbox. After some investigation it turns out that returning
the response to the top of the command handler is not required.