fix(GODT-1570): Fix too many SQL Variable in sync benchmark
Batch incoming connector updates for new message to 1000 messages per
transaction to avoid running into too many SQL variables.

We batched at outside of the transaction level so as not to completely
block the other threads forever until we are done processing a high
number of updates.

The code has also been updated to use a more optimized database query.