refactor(GODT-1570): Reuse parsed contents for messages
Update the `Connector.CreateMessage` to also require a parsed message
type. This enabled the connector implementation to inspect the message
without having to re-parse the message as we already do this before the
connector request.

This patch separates the parsed message from the `MessageCreated` update
so that we can use it in other locations in the code base.