refactor(GODT-1730): Rewrite Header Parser
Rewrite the header parser so that it can be used to sequentially query
all the available header items as we go.

Updated `GetHeaderValue()` to use the new parser, we can now check if
the next header item matches the key we are looking for, rather than
parsing the whole header first and then performing a lookup.

Finally this patch also fixes one invalid test (folding headers always
need to start with space or tab) and adds a new test to make sure the
prelude is handled properly.