refactor(GODT-1730): Refactor HeaderType
To improve lookup after the header has been parsed all the entries are
stored in a map.

Every single line of the header is stored as a linked list so we
can perform order dependent queries or modify the header while
preserving the original order. The result of mergedLine requests are
cached to avoid this computation in the future.

The SetHeaderValue function has been updated so that it does not parse
the full header, but rather sets the new value as the first value of the
new header. This avoids having to parse the whole header to set our
internal ID-Key

Finally our internal ID key was case-corrected according to the
standard.