fix: Header parser (naive)
The new header parser did not correctly deal with two edge cases:
1. The message contains an embedded RFC822 message with no header
2. Parsing something (not necessarily a header) with a key but no value

This fix checks that what we think is a header actually is one in a
naive way by trying to parse it and seeing if it returns an error.
It also handles the case that a header key was not finished being read
by returning an error, indicating this wasn't a valid header).