refactor(GODT-1570): Refactor flag handling code
Use snap message slices directly, avoids doing redundant calculations
to retrieve flags from the snapshot.

Avoid using `xslices.Filter` so we don't have to maintain a map for
lookup.

Do not return final flag map as it is never used and only call to remote
if there changes that need to be made.