refactor(GODT-1570): Do not track Seq ID in snapshot messages
This change improves the performance of expunged ranges by approximately
60% since we do no longer have to keep updating all the indices for the
remaining items and their lookup up maps.

The Seq ID is now generated on demand based of the location of the
message in the message list.

This changes only adds a minor increase to lookup times for other
operations, usually in the 20-30 micro second range.