refactor: Resurrect transaction-free on-disk store
Badger imposes some limitations to what we can do with it in terms of
transactions. We originally needed transaction support to safely perform
renames of files on disk due to remote ID changes. This operation no
longer exists and we accept some stale data in the on disk cache until
the cleanup routine is run on startup.