fix(GODT-1570): Handle literals in IMAP Parser without map + C API
Adjust the parser so that we don't have to pass in a literal map to
parse the message. We removed the Swig API in favor of a simpler C style
API and commands are no longer serialized to string but straight to a
byte array.

This patch also changes the representation of the delimiter as it is
supposed to be a single character rather than a string.