refactor: Return newly created message literal on import
It's possible for the proton API to modify the imported message.
To ensure users see the same thing locally as they do on the API,
we make the connector return the server-side message literal after
import. The returned message literal is then the one added to the
gluon database, rather than the client-created one.