feat: MessageUpdated Update
To be used in cases where one needs to atomically replace a message
(e.g: Drafts) in one go. This was previously possible with a combination
of Delete + Create update, but these were not guaranteed to be atomic
and could cause undefined behaviors if something else happened
in between.

The old logic to handle the replacement of delete messages on create has
been reverted. It is now expected that this type of behavior uses the
newly introduced update.