feat(GODT-1989): Recovery Mailbox
Add a Gluon managed mailbox which stores messages which failed to be
appended. This is required to handle clients such as Outlook which do
not check Append failures when moving message and can result in data
loss.

When an Append fails, we insert it into a special mailbox called
"RecoveredMessages". Messages in this mailbox support:
* Being copied out
* Being moved out
* Flag modifications
* Being deleted

Any other operation is not allowed.

When messages are moved/copied out of this mailbox they are first
created as new messages on the Connector before being added to the
target mailbox.