fix(GODT-2170): RemoveMessagesFromMailbox should only remove from remote if need be
When adding a message to a mailbox, it can happen that the message is
already there. In that case, we remove it and re-add it. However, that
change only needs to be done locally; it's useless to do it on the
remote. Further, it can even be dangerous to do on the remote, as
the remote might handle this as a permanent deletion.