fix(GODT-2170): Ensure Append into Drafts always creates messages
To better handle the workflows of editing Drafts on IMAP clients, we now
always force Append in Drafts to create a new message. We also remove
previous internal IDs to avoid confusion.

This currently implies that the connector can't return the same Remote
ID for this message. We now fail and report an exception when we detect
this scenario.