fix(GODT-1896): Do not remove subscribed status for deleted mailboxes
Every time we delete a mailbox, if this mailbox is subscribed it is
stored in the `DeletedSubscriptions` table. If we create a new mailbox
with the same name, the entry is removed from the table.

This approach was chose as it does not require any special database
migrations other than creating the new table.