fix(GODT-2188): Do not fail APPEND  with invalid MIME types
If we run into an invalid mime type do not block parsing and upload of
the message. When this happen we simply report nothing as MIME type.
This works as we only use this information internally to reply to FETCH
BODY and BODYSTRUCTURE requests.

Finally we also ensure that the Content-Type value only contains valid
ASCII characters before being processed.