feat(GODT-2252): Recover from missing cache files
Allow commands such as fetch and search to re-download literals from the
connector if they are no longer available.

To ensure parallel fetches which download the same literal don't cause
file system issue with writing multiple files, we wrap the store
implementation with a layer that controls access to the message with a
RWLock.