refactor(GODT-2224): Refactor store backend
Refactor store to use less memory when reading and writing files and
swap the compression algorithm to LZ4 for faster decompression.

Rather than compressing the whole file and then encrypted all of the
compressed data, we now encrypt `BlockSize` compressed data and then
write it to disk. This way we only need to keep +/- 2x BlockSize in
memory for reading/writing.

Note: This is a breaking change as the store format has now changed.