feat(GODT-2201): Native GO IMAP Parser Foundations
Lay the foundations for native parsers in Go so we can move away from
the C++ libraries.

This patch sets up the foundations to convert the remaining commands to
this new format. At the moment the List and Append command have been
ported and should serve as an example for the remaining commands.

More commands will follow in future patches.