fix(GODT-2399): Defer updated message deletion
Ensure that when a message gets update, we defer deletion of the message
so that it remains accessible until all snapshots that have this message
have been released.

Note: To work around not migrating the internal database, when updating
a message, the internal ID will be replaced with `DELETED-<new uuid>` so
that the table constraints do not fail.