fix(GODT-2526): Unlimited memory usage during fetch/search
This bug was introduced when we switch for a queued channel to handle
the client's responses. This had the unfortunate side effect that if the
responses would arrive faster than they can be sent to client they can
pile up. When this is combined with fetch queries that read the message
literal, an unhealthy amount of RAM can be consumed by this.