fix(GODT-2573): Remove Prelude header parsing support
The handling of Preludes in the headers stem from the MBox format.
According to the IMAP spec, the literal submitted to APPEND should be a
valid RFC822 formatted message.

The handling of the prelude was also causing issues where an MBox
formatted entry would slip through and then cause issues later.