fix(GODT-2454): Only apply state updates if db transactions succeed
Ensure that all changes to the DB are made before publishing state
updates for connected sessions. This improves stability of the connected
clients.

The states still require another transaction to be applied. However,
this is currently only related to the handling of the recent flag. This
will be handled in GODT-2522 instead.