feat(GODT-2510): Direct SQLite3
DB implementation using SQLite3. This current version is feature
complete with the ent implementation. The only noticeable difference is
the addition of the db version table.

The new database is now the default, but we will still run tests on CI
to compare behavior against old implementation until it is completely
removed.