feat(GODT-2803): Allow access to IMAP state from connector
Update connector interface so that it has access to the internal IMAP
state of Gluon.

Similar to the DB design, there are modifiers for writing and reading
from this state. This is required so that it remains compatible with
how we are currently handling the database access code.

Finally, The `StateConnectorInterface` has been updated to return state
updates for all actions, since it is now possible for connectors to
apply IMAP state changes with each operation. Similar to previous
changes, those state updates need to be kept around until the
transaction has successfully committed.

This patch includes 3 example functions which cover the cases of
reading, writing without updates and writing with updates. More should
be added in the future as needed.