feat(GODT-2576): Support for Forward flags
Necessary changes required to support forward flags. These are
non-standard IMAP flags that can be assigned to mailboxes. Thunderbird
uses `$Forwarded` and Apple Mail issues both `$Forwarded` and `Forwarded`.
Unfortunately, Outlook (office 365) does not support any non-standard
IMAP flags.

This patch extends the connector with the `MarkMessageForwarded` as the
forwarded setting is applied after a message is sent.

Every time we apply a forward flag, we ensure that all known forwarding
flags are set so that it works with all known variations.